Struct hex2d_dpcext::algo::bfs::Traverser

pub struct Traverser<FCanPass, FIsDest, I = i32> where I: Integer, I: Hash, FCanPass: Fn(Coordinate<I>) -> bool, FIsDest: Fn(Coordinate<I>) -> bool {
    // some fields omitted
}

Breadth First Search

Use BFS to find closest (in walk steps) Coordinates that satisfy is_dest and can be reached with a walk through coordinates for which can_pass returns true.

Methods

impl<FCanPass, FIsDest, I> Traverser<FCanPass, FIsDest, I> where I: Integer, I: Hash, FCanPass: Fn(Coordinate<I>) -> bool, FIsDest: Fn(Coordinate<I>) -> bool

fn new(can_pass: FCanPass, is_dest: FIsDest, start: Coordinate<I>) -> Traverser<FCanPass, FIsDest, I>

Create a Traverser instance with initial conditions

fn find(&mut self) -> Option<Coordinate<I>>

Find next closest coordinate.

Can be called multiple times, each time returning next coordinate

fn backtrace(&self, pos: Coordinate<I>) -> Option<Coordinate<I>>

Return neighbor Coordinate to pos that is one step closer to start from initial conditions.

Useful for finding whole path to a Coordinate returned by find.

Returns None for Coordinates that were not yet visited. Returns start for start (from initial conditions)

fn backtrace_last(&self, pos: Coordinate<I>) -> Option<Coordinate<I>>

Perform a recursive backtrace walk to find a neighbor of start that leads to the Coordinate returned by find().

Returns None for Coordinates that were not yet visited. Returns start for start (from initial conditions)