Function hex2d_dpcext::algo::los2::los

pub fn los<FOpaqueness, FVisible, I = i32>(opaqueness: &FOpaqueness, visible: &mut FVisible, light: I, pos: Coordinate<I>, dirs: &[Direction]) where I: Integer, I: Hash, &'a I: Add<&'a I, Output=I>, FOpaqueness: Fn(Coordinate<I>) -> I, FVisible: FnMut(Coordinate<I>, I)

Starting from pos in direction dir, call visible for each visible Coordinate.

Use light as starting range of the LoS. For each visible Coordinate, the value returned by opaqueness will be subtracted from light to check if the LoS should finish due to "lack of visibility". opaqueness should typically return 1 for fully transparent Coordinates, and anything bigger than initial light for fully opaque Coordinates.